address first round of Paul comments
