+++ index.html
@@ -14,6 +14,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/knapsack/main.tsx"></script>
+    <script type="module" src="/src/DocExplorer/main.tsx"></script>
   </body>
 </html>
