+++ standalone.html
@@ -1,4 +1,4 @@
-<!-- an entrypoint file that just renders TEE, with no Knapsack frame -->
+<!-- an entrypoint file that just renders TEE, with no Explorer frame -->
 
 <!DOCTYPE html>
 <html lang="en">
